move AzureServiceClient to fluentcore
