+++ sdk/authorization/mgmt/src/main/java/com/azure/resourcemanager/authorization/GraphRbacManagementClient.java
@@ -8,7 +8,7 @@ import com.azure.core.annotation.ServiceClient;
 import com.azure.core.http.HttpPipeline;
 import com.azure.core.management.AzureEnvironment;
 import com.azure.core.util.logging.ClientLogger;
-import com.azure.resourcemanager.AzureServiceClient;
+import com.azure.resourcemanager.resources.fluentcore.AzureServiceClient;
 import com.azure.resourcemanager.authorization.fluent.ApplicationsClient;
 import com.azure.resourcemanager.authorization.fluent.DeletedApplicationsClient;
 import com.azure.resourcemanager.authorization.fluent.DomainsClient;
