--- sdk/resources/mgmt/src/main/java/com/azure/resourcemanager/package-info.java
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for
-// license information.
-
-/**
- * This package contains the base classes for management sdk.
- */
-package com.azure.resourcemanager;
