+++ sdk/resources/mgmt/src/main/java/com/azure/resourcemanager/resources/fluentcore/AzureServiceClient.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.resourcemanager;
+package com.azure.resourcemanager.resources.fluentcore;
 
 import com.azure.core.http.HttpPipeline;
 import com.azure.core.http.rest.Response;
@@ -28,20 +28,15 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import javax.xml.bind.annotation.adapters.HexBinaryAdapter;
 import java.io.IOException;
 import java.lang.reflect.Type;
-import java.net.NetworkInterface;
 import java.nio.ByteBuffer;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.TemporalAccessor;
 import java.time.temporal.TemporalQueries;
-import java.util.Enumeration;
 
 /**
  * ServiceClient is the abstraction for accessing REST operations and their payload data types.
@@ -54,20 +49,6 @@ public abstract class AzureServiceClient {
         ((AzureJacksonAdapter) serializerAdapter).serializer().registerModule(DateTimeDeserializer.getModule());
     }
 
-    /**
-     * The default User-Agent header. Override this method to override the user agent.
-     *
-     * @return the user agent string.
-     */
-    public String userAgent() {
-        return String.format("Azure-SDK-For-Java/%s OS:%s MacAddressHash:%s Java:%s",
-            getClass().getPackage().getImplementationVersion(),
-            OS,
-            MAC_ADDRESS_HASH,
-            JAVA_VERSION);
-    }
-
-    private static final String MAC_ADDRESS_HASH;
     private static final String OS;
     private static final String OS_NAME;
     private static final String OS_VERSION;
@@ -78,22 +59,6 @@ public abstract class AzureServiceClient {
         OS_NAME = System.getProperty("os.name");
         OS_VERSION = System.getProperty("os.version");
         OS = OS_NAME + "/" + OS_VERSION;
-        String macAddress = "Unknown";
-        try {
-            Enumeration<NetworkInterface> networks = NetworkInterface.getNetworkInterfaces();
-            while (networks.hasMoreElements()) {
-                NetworkInterface network = networks.nextElement();
-                byte[] mac = network.getHardwareAddress();
-
-                if (mac != null) {
-                    macAddress = getSha256(mac);
-                    break;
-                }
-            }
-        } catch (Throwable t) {
-            // It's okay ignore mac address hash telemetry
-        }
-        MAC_ADDRESS_HASH = macAddress;
         String version = System.getProperty("java.version");
         JAVA_VERSION = version != null ? version : "Unknown";
     }
@@ -148,7 +113,7 @@ public abstract class AzureServiceClient {
      * @param <U> type of final result.
      * @return poller flux for poll result and final result.
      */
-    public <T, U> PollerFlux<PollResult<T>, U> getLroResultAsync(Mono<? extends Response<Flux<ByteBuffer>>> lroInit,
+    public <T, U> PollerFlux<PollResult<T>, U> getLroResultAsync(Mono<Response<Flux<ByteBuffer>>> lroInit,
                                                                  HttpPipeline httpPipeline,
                                                                  Type pollResultType, Type finalResultType) {
         return PollerFactory.create(
@@ -157,23 +122,8 @@ public abstract class AzureServiceClient {
             pollResultType,
             finalResultType,
             SdkContext.getLroRetryDuration(),
-            activationOperation(lroInit)
+            lroInit
         );
-    }
-
-    private Mono<Response<Flux<ByteBuffer>>> activationOperation(Mono<? extends Response<Flux<ByteBuffer>>> lroInit) {
-        return lroInit.flatMap(fluxSimpleResponse -> Mono.just(fluxSimpleResponse));
-    }
-
-    private static String getSha256(byte[] bytes) {
-        MessageDigest messageDigest;
-        try {
-            messageDigest = MessageDigest.getInstance("SHA-256");
-            return new HexBinaryAdapter().marshal(messageDigest.digest(bytes));
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-        return "Unknown";
     }
 
     /**
