+++ sdk/resources/mgmt/src/main/java/com/azure/resourcemanager/resources/fluentcore/package-info.java
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * This package contains the base classes for management sdk.
+ */
+package com.azure.resourcemanager.resources.fluentcore;
